Update to debuerreotype 0.12

+++ tianon/debuerreotype/arch-pipeline.groovy
@@ -42,10 +42,12 @@ node(multiarchVars.node(env.BUILD_ARCH, env.ACT_ON_IMAGE)) {
 			stage('Download') {
 				sh '''
 					wget -O 'debuerreotype.tgz' "https://github.com/debuerreotype/debuerreotype/archive/${debuerreotypeVersion}.tar.gz"
-					wget -O 'debuerreotype-examples.tgz' "https://github.com/debuerreotype/debuerreotype/archive/${debuerreotypeExamplesCommit}.tar.gz"
 					tar -xf debuerreotype.tgz --strip-components=1
-					rm -rf examples
-					tar -xf debuerreotype-examples.tgz --strip-components=1 "debuerreotype-${debuerreotypeExamplesCommit}/examples"
+					if [ "$debuerreotypeExamplesCommit" != "$debuerreotypeVersion" ]; then
+						wget -O 'debuerreotype-examples.tgz' "https://github.com/debuerreotype/debuerreotype/archive/${debuerreotypeExamplesCommit}.tar.gz"
+						rm -rf examples
+						tar -xf debuerreotype-examples.tgz --strip-components=1 "debuerreotype-${debuerreotypeExamplesCommit}/examples"
+					fi
 					rm -f debuerreotype*.tgz
 					./scripts/debuerreotype-version
